Один тип плагина для bsl и md.
