+++ plugins/__init__.py
@@ -1,0 +1,15 @@
+
+from abc import ABC, abstractmethod
+
+class PluginResult:
+    pass
+
+class Plugin(ABC):
+
+    @abstractmethod
+    def __init__(self, path: str, src: str):
+        pass
+
+    @abstractmethod
+    def close(self) -> PluginResult:
+        pass
